backend/requirements.txt
astroid==2.2.5
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.11
django-allauth==0.39.1
django-cors-headers==2.5.2
django-rest-auth==0.9.3
djangorestframework==3.9.2
idna==2.8
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
Pillow==5.4.1
psycopg2==2.7.7
pycodestyle==2.5.0
pylint==2.3.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
